backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
alembic==1.13.1
redis==5.0.6
openai==1.40.0
python-dotenv==1.0.1
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4
scipy==1.13.1
essentia==2.1b6.dev609
keyfinder==2.2.7
aubio==0.4.9
google-search-results==2.4.2






